+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * ADD: `File.birthtime` and `File#birthtime`
 * ADD: `File.empty?`
 * ADD: `File::Stat#nlink` (#39 by @djberg96)
+* ADD: `Dir.mktmpdir` (#52 by @djberg96)
 * FIX: Fixing the inverted _read_ and _execute_ bitmasks (#41 by @micahlee)
 * ADD: Dependabot configuration
 * CHG: Replacing Travis CI with GitHub Actions